Added existance check on migrations

+++ database/migrations/2024_01_16_211945_add_created_by_id_to_project_summaries_table.php
@@ -12,10 +12,13 @@ return new class extends Migration
     public function up()
     {
         Schema::table('project_summaries', function (Blueprint $table) {
-            $table->foreignId('createdById')->constrained('users');
+            // Check if the column doesn't exist before adding
+            if (!Schema::hasColumn('project_summaries', 'createdById')) {
+                $table->unsignedBigInteger('createdById')->nullable(); 
+                $table->foreign('createdById')->references('id')->on('users');
+            }
         });
     }
-
 
     /**
      * Reverse the migrations.
@@ -23,8 +26,12 @@ return new class extends Migration
     public function down()
     {
         Schema::table('project_summaries', function (Blueprint $table) {
-            $table->dropForeign(['createdById']);
-            $table->dropColumn('createdById');
+            // Check if the foreign key exists before dropping
+            if (Schema::hasColumn('project_summaries', 'createdById') && 
+                Schema::hasForeignKey('project_summaries', 'project_summaries_createdById_foreign')) {
+                $table->dropForeign(['createdById']);
+                $table->dropColumn('createdById');
+            }
         });
     }
