+++ database/migrations/2024_01_16_212027_add_created_by_id_to_meeting_summaries_table.php
@@ -9,21 +9,29 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up(): void
+    public function up()
     {
         Schema::table('meeting_summeries', function (Blueprint $table) {
-            $table->foreignId('createdById')->constrained('users');
+            // Check if the column doesn't exist before adding
+            if (!Schema::hasColumn('meeting_summeries', 'createdById')) {
+                $table->unsignedBigInteger('createdById')->nullable(); 
+                $table->foreign('createdById')->references('id')->on('users');
+            }
         });
     }
 
     /**
      * Reverse the migrations.
      */
-    public function down(): void
+    public function down()
     {
         Schema::table('meeting_summeries', function (Blueprint $table) {
-            $table->dropForeign(['createdById']);
-            $table->dropColumn('createdById');
+            // Check if the foreign key exists before dropping
+            if (Schema::hasColumn('meeting_summeries', 'createdById') && 
+                Schema::hasForeignKey('meeting_summeries', 'meeting_summeries_createdById_foreign')) {
+                $table->dropForeign(['createdById']);
+                $table->dropColumn('createdById');
+            }
         });
     }
 };
